requirements.txt
Django==4.2
djangorestframework==3.16.0
pillow==11.1.0
psycopg2-binary==2.9.10
psycopg2
python-dotenv==1.1.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10